Sync: fail-closed frame parsing + strict version checks (SEC-006)
Both platforms now reject malformed sync frames before they can mutate
state. Adds strictVersion (rejects NaN, Infinity, fractional, negative,
boolean, out-of-range, wrong JSON type) and frame-level ceilings (1 MiB
text, 1 MiB Base64, 10K snapshot items, MAX_V=1e12). iOS versions/clock
widened from Int to Int64. Clock advancement moved after AEAD+sig verify
to prevent poisoning by a hostile relay. Shared test corpus in
testdata/malicious_frames.json (25 vectors) runs on both platforms.

+++ ios/TacticalMaps/Sync/SyncCrypto.swift
@@ -90,7 +90,7 @@ enum SyncCrypto {
     }
 
     /// AEAD associated data - binds ciphertext to routing metadata.
-    static func aad(id: String, v: Int, kind: String) -> Data {
+    static func aad(id: String, v: Int64, kind: String) -> Data {
         Data("\(id)|\(v)|\(kind)".utf8)
     }
